debugging

+++ ft_abs.c
@@ -1,6 +1,6 @@
 #include "ft_printf.h"
 
-int	ft_abs(long int a)
+long int	ft_abs(long int a)
 {
 	if (a < 0)
 		return (a * -1);
